[CRNT-37915]:Benchmark Management: Deleting Benchmark Condition Category should not be allowed when there is only one Category present.
Modified so that if user attempts to delete the last category then they are presented with a dialog "Benchmark definitions must contain at least one category" and the delete operation is circumvented.

Bug: CRNT-37915

+++ benchmark/src/main/resources/META-INF/webapp/html5/scripts/controllers/benchmarkController.js
@@ -198,6 +198,7 @@
 		this.textMap.dialogButtonCancel  = this.i18N("views.main.dialog.buttons.cancel.default");
 		this.textMap.dialogButtonCancelOk  = this.i18N("views.main.dialog.buttons.cancel.ok");
 		this.textMap.dialogButtonCancelClose  = this.i18N("views.main.dialog.buttons.cancel.close");
+		this.textMap.categoryDeleteDeny = this.i18N("views.main.dialog.category.delete.deny");
 	};
 	
 	/**
@@ -733,7 +734,13 @@
 			benchmark.isDirty = true;//v.item.bm.isDirty = true;
 			switch (v.item.action){
 				case "KILL_CAT":
-					this.benchmarkBuilderService.removeCategory(benchmark,category);
+					if(categories.length===1){
+						this.categoryDeleteDeny.open();
+					}
+					else{
+						this.benchmarkBuilderService.removeCategory(benchmark,category);
+					}
+					
 					break;
 				case "ADD_CAT":
 					this.benchmarkBuilderService.addCategory(benchmark,category);
